+++ aspnet-core/src/Fanap.DataLabeling.Application/Credit/GetCreditOutputDto.cs
@@ -4,7 +4,7 @@ namespace Fanap.DataLabeling.Credit
 {
     public class GetCreditOutput
     {
-        public decimal Credit { get; internal set; }
+        public double Credit { get; internal set; }
         public TargetDefinitionDto Target { get; internal set; }
     }
 }
